feat(step1): 계산식 단일 숫자 포장
