+++ src/test/java/com/kkambi/calculator/domain/NumbersTest.java
@@ -2,13 +2,12 @@ package com.kkambi.calculator.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
 
 import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Collectors;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 
 class NumbersTest {
 
@@ -16,34 +15,16 @@ class NumbersTest {
     @Test
     void createNumbersWithNumberList() {
         // given
-        String[] elements = {"1", "2", "3"};
+        String[] tokens = {"1", "2", "3"};
+        Elements elements = new Elements(tokens);
 
         // when
-        assertThat(new Numbers(elements).getNumbers())
-                .containsExactlyElementsOf(Arrays.asList(1, 2, 3));
-    }
+        List<Number> numbers = new Numbers(elements).getNumbers();
+        List<Integer> result = numbers.stream()
+                .map(Number::getNumber)
+                .collect(Collectors.toList());
 
-    @DisplayName("숫자가 아닌 값을 받으면 예외를 던진다")
-    @ValueSource(strings = {"*", "k", "A", ","})
-    @ParameterizedTest
-    void throwExceptionWhenNotNumber(String element) {
-        // given
-        String[] elements = {element};
-
-        // when
         // then
-        assertThrows(RuntimeException.class, () -> new Numbers(elements));
-    }
-
-    @DisplayName("음수를 받으면 예외를 던진다")
-    @ValueSource(strings = {"-1", "-10", "-999"})
-    @ParameterizedTest
-    void throwExceptionWhenNegative(String element) {
-        // given
-        String[] elements = {element};
-
-        // when
-        // then
-        assertThrows(RuntimeException.class, () -> new Numbers(elements));
+        assertThat(result).containsExactlyElementsOf(Arrays.asList(1, 2, 3));
     }
 }
